feat: change the MapPattern raw characters to an enum, store the pattern in the world

+++ api/indiestudio/world/WorldSettings.hpp
@@ -31,8 +31,8 @@ namespace IndieStudio {
         };
 
         std::string name;
-        unsigned int width = 17;
-        unsigned int height = 11;
+        unsigned short width = 17;
+        unsigned short height = 11;
         std::string worldGenerator = "Basic";
         Player players[4];
